genai_application/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
clear==2.0.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
confluent-kafka==2.4.0
cymem==2.0.8
dnspython==2.6.1
email_validator==2.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.33.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.54.0
grpcio-status==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.24.1
orjson==3.10.5
packaging==24.1
portalocker==2.10.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.16
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.1
qdrant-client==1.9.2
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
thinc==8.2.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
wasabi==0.10.1
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0